BigKurius/requirements.txt
discord.py>=1.6.0
Flask>=2.0.2